+++ apps/docs/src/content/03-components/chat/message-separator/examples/default.tsx
@@ -1,3 +1,0 @@
-import { MessageSeparator } from "@mittwald/flow-react-components";
-
-<MessageSeparator>Ticket erstellt</MessageSeparator>;