+++ apps/docs/src/content/03-components/structure/list/examples/sorting.tsx
@@ -15,39 +15,29 @@ import {
   Text,
   typedList,
 } from "@mittwald/flow-react-components";
-import { DateTime } from "luxon";
 
 export default () => {
   type DomainWithBigIntId = Omit<Domain, "id"> & {
     id: bigint;
-    createdAt: DateTime;
   };
 
-  const domainsWithDateTime = domains.map(
-    (domain, index) => {
-      const daysAgo =
-        index * 3 + Math.floor(Math.random() * 5);
+  const domainsWithBigInt = domains.map((domain, index) => {
+    const bigIntId = BigInt(1000000000000 + index);
 
-      const bigIntId = BigInt(1000000000000 + index);
-
-      return {
-        ...domain,
-        id: bigIntId,
-        createdAt: DateTime.now().minus({ days: daysAgo }),
-      };
-    },
-  );
+    return {
+      ...domain,
+      id: bigIntId,
+    };
+  });
 
   const DomainList = typedList<DomainWithBigIntId>();
 
   const bigIntSorting =
     SortingFunctions.bigInt as SortingFn<DomainWithBigIntId>;
-  const dateTimeSorting =
-    SortingFunctions.dateTime as SortingFn<DomainWithBigIntId>;
 
   return (
     <DomainList.List batchSize={5} aria-label="Domains">
-      <DomainList.StaticData data={domainsWithDateTime} />
+      <DomainList.StaticData data={domainsWithBigInt} />
 
       <DomainList.Sorting
         property="hostname"
@@ -72,19 +62,6 @@ export default () => {
         direction="desc"
         customSortingFn={bigIntSorting}
         defaultEnabled
-      />
-
-      <DomainList.Sorting
-        property="createdAt"
-        name="Erstellt am (älteste zuerst)"
-        direction="asc"
-        customSortingFn={dateTimeSorting}
-      />
-      <DomainList.Sorting
-        property="createdAt"
-        name="Erstellt am (neueste zuerst)"
-        direction="desc"
-        customSortingFn={dateTimeSorting}
       />
 
       <DomainList.Sorting
@@ -127,10 +104,6 @@ export default () => {
             <Text>{domain.type}</Text>
             <Text>ID: {domain.id}</Text>
             <Text>TLD: {domain.tld}</Text>
-            <Text>
-              Erstellt am:{" "}
-              {domain.createdAt.toLocaleString()}
-            </Text>
 
             <ContextMenu>
               <MenuItem>Details anzeigen</MenuItem>