+++ packages/components/src/components/CartesianChart/components/XAxis/XAxis.tsx
@@ -16,6 +16,7 @@ export type XAxisProps = Pick<
   | "domain"
   | "hide"
   | "unit"
+  | "tickFormatter"
 >;
 
 /** @flr-generate all */