+++ packages/components/src/components/MessageThread/components/MessageSeparator/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./MessageSeparator";
+export * from "./MessageSeparator";