+++ packages/components/src/components/PasswordCreationField/PasswordCreationField.tsx
@@ -181,7 +181,6 @@ export const PasswordCreationField = flowComponent(
         setOptimisticPolicyValidationResult({
           isValid: "indeterminate",
         });
-        onChangeHandler(pastedValue);
       }
     };