+++ packages/remote-elements/src/auto-generated/RemoteXAxisElement.ts
@@ -21,6 +21,7 @@ export class RemoteXAxisElement extends FlowRemoteElement<RemoteXAxisElementProp
       minTickGap: {},
       orientation: {},
       scale: {},
+      tickFormatter: {},
       type: {},
       unit: {},
     };