+++ packages/remote-elements/src/auto-generated/RemoteYAxisElement.ts
@@ -20,6 +20,7 @@ export class RemoteYAxisElement extends FlowRemoteElement<RemoteYAxisElementProp
       minTickGap: {},
       orientation: {},
       scale: {},
+      tickFormatter: {},
       type: {},
       unit: {},
     };